Update GUI.java
You can see that when serializing the Password object in the main method, I will continue to use the same name of the file.  What I hoped was to find a way how to let the program know that the file should be serialized once for the next time a user would like to log in to his/her account.  I realize that a database would be best for my project's main function, but I figured I would try to implement serializing as mini-version of a more complex database.